requirement.txt
fastapi
uvicorn
sqlalchemy
pydantic
databases
python-dotenv
passlib[bcrypt]

